pyproject.toml
[tool.poetry]
name = "cam-server"
version = "0.1.0"
description = ""
authors = ["josericardopenase <josericardopenase@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.9.0.80"
flask = "^3.0.0"
peewee = "^3.17.0"
streams-py = "^1.2.0"
flask-cors = "^4.0.0"
flask-socketio = "^5.3.6"
mediapipe = "^0.10.9"
pillow = "^10.2.0"
aiohttp = "^3.9.1"
matplotlib = "^3.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "main:start"